requirements.txt
nnunet==1.6.4
git+https://github.com/FabianIsensee/hiddenlayer.git@more_plotted_details#egg=hiddenlayer
batchgenerators==0.21
tensorflow-gpu==2.4.1
tensorflow_addons==0.11.2
glob2==0.7
h5py==2.10.0
Keras==2.4.3
matplotlib==3.3.4
nibabel==2.5.1
numpy==1.19.5
#pandas==1.2.4
#scipy==1.6.2
#sklearn==0.0
#scikit-build==0.12.0
#SimpleITK==1.2.0
#nilearn==0.6.0
#nipype==1.3.1

